Add communication with the snapshot-processor
Add fetch and form data methods to communicate with the snapshot
processor. Add plugin file due to requirements of the monkey by Juan.
Add code to the plugin to manage the lifecycle of the snapshots.
Minor changes to the new verison of the Juan's monkey.
Modify package.json to include new dependencies for the communication.